requirements.txt
ansitable==0.11.4
asttokens==3.0.0
cfgv==3.4.0
contourpy==1.3.1
colored==2.3.0
cycler==0.12.1
dateutils==0.6.12
decorator==5.1.1
distlib==0.3.9
executing==2.2.0
fonttools==4.56.0
filelock==3.18.0
iniconfig>=1.1.1
identify==2.6.10
ipdb==0.13.13
ipython==8.32.0
jedi==0.19.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
numpy==1.24.3
narwhals==1.27.1
nodeenv==1.9.1
opencv-python==4.8.0.74
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pgraph-python==0.6.3
pillow==11.1.0
pluggy>=1.0.0
plotly==6.0.0
pre-commit==4.2.0
prompt-toolkit==3.0.50
progress==1.6
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.1
pytest>=7.0.0
python_dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
rich==13.9.4
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.14.1
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.14
swift-sim==1.1.0
stack-data==0.6.3
traitlets==5.14.3
typing-extensions==4.12.2
virtualenv==20.31.1
wcwidth==0.2.13
websockets==15.0.1
